Add the ability to administratively enroll repo into using superproject.
Repo will remember a choice and an expiration time of the choice, per
user, about whether to use superproject by default. When not specified
from command line and the choice is not expired, repo would use the
user default value.

When a user default value is not present and when the system wide
enable default is provided at /usr/share/git-repo/config, repo would
ask the user for a confirmation which will be valid for two weeks.

git_config.py: Add support for /usr/share/git-repo-config

git_superproject.py: Add a new subroutine, _UseSuperproject(), which
	returns whether superproject should be used and whether it
	is from a user configuration.

	The value is determined in the following order:

	1. If the user specifies either --use-superproject or
	--no-use-superproject, then that choice is being used.

	2. If neither is specified, we would then check the saved value
	(upon repo init) and use that choice when there was a choice.

	3. We then check if there is a saved and unexpired value for
	user's choice in their ~/.gitconfig, and use the unexpired
	choice, if available.

	4. Finally, if all the above didn't give us a decision, and if
	/usr/share/git-repo-config is providing a rollout hint, present
	a prompt to user for their decision and save it in ~/.gitconfig.

subcmds/sync.py: Make use of the new UseSuperproject() provided by
git_superproject.py.

While there also silent stderr from git describe when determining the
version of repo.

+++ git_config.py
@@ -64,6 +64,7 @@ class GitConfig(object):
   _ForUser = None
 
   _USER_CONFIG = '~/.gitconfig'
+  _SYSTEM_CONFIG = '/usr/share/git-core/config'
 
   @classmethod
   def ForUser(cls):
@@ -76,9 +77,16 @@ class GitConfig(object):
     return cls(configfile=os.path.join(gitdir, 'config'),
                defaults=defaults)
 
-  def __init__(self, configfile, defaults=None, jsonFile=None):
+  @classmethod
+  def ForSystem(cls):
+    if cls._ForSystem is None:
+      cls._ForSystem = cls(configfile=cls._SYSTEM_CONFIG, readonly=True)
+    return cls._ForSystem
+
+  def __init__(self, configfile, defaults=None, jsonFile=None, readonly=False):
     self.file = configfile
     self.defaults = defaults
+    self.readonly = readonly
     self._cache_dict = None
     self._section_dict = None
     self._remotes = {}
@@ -204,6 +212,9 @@ class GitConfig(object):
        The supplied value should be either a string,
        or a list of strings (to store multiple values).
     """
+    if self.readonly:
+      raise AttributeError("Can't write read-only configuration")
+
     key = _key(name)
 
     try:
@@ -323,6 +334,9 @@ class GitConfig(object):
       return None
 
   def _SaveJson(self, cache):
+    if self.readonly:
+      return
+
     try:
       with open(self._json, 'w') as fd:
         json.dump(cache, fd, indent=2)
@@ -372,7 +386,9 @@ class GitConfig(object):
 class RepoConfig(GitConfig):
   """User settings for repo itself."""
 
+  _ForSystem = None
   _USER_CONFIG = '~/.repoconfig/config'
+  _SYSTEM_CONFIG = '/usr/share/git-repo/config'
 
 
 class RefSpec(object):
